requirements.txt
pandas
lxml
tqdm
gensim
medspacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz
boto3
awswrangler
python-dotenv
sqlalchemy
sqlalchemy-cockroachdb
psycopg2

